pyproject.toml
[project]
name = "trading-agents"
version = "0.1.0"
description = "Self-hosted multi-agent system for ML-based trading"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "tensorflow>=2.15.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.36",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "mlflow>=2.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.3.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"